add TilesableData, make TensorData inherit from it
